requirements.txt
Flask==3.0.3
botbuilder-core==4.17.0
